refactor(slideshow): remove unused FormComboboxField and FormSelectField components
- Deleted FormComboboxField and FormSelectField components to streamline the codebase.
- Updated index file to reflect the removal and adjust imports accordingly.
- Refactored StepEducation component to utilize updated imports for better organization.
